Changed user mail to name
Simple user that are created by the main User should not need to have an
email. It should be possible to create accounts for People without email.
For example Kids.
Email should be required to use the App in the end.
For example for somebody who want to synchronize more than one device or
connect with family member