Slider in activity and nutrition setting will now be colored
If the sum of the entries is above the available value,
all sliders turn red, and if the sum is below the available value, the sliders turn yellow.
If the value is exactly in the available range, all sliders turn Green
A click on the respective icon, matching in the slider map, push the value to the required difference.